+++ app/src/main/java/activity/PermissionValidator.java
